fix(ios): complete grade canonicalization for clean compile
Windows-side commit 113f091 introduced gradeCompany/gradeValue to
InventoryCard but missed several call sites + had Int? vs Double?
type mismatch with backend wire format (gradeValue: number, which can
be 9.5 for BGS grades).

Mac-side compile fixes:

1. APIService.swift: CardEstimateRequest.gradeValue Int? to Double?
   to match backend wire format.

2. CompatibilityShims.swift: Added gradeCompany/gradeValue to
   CodingKeys + SnakeKeys + decoder + encoder + updatingCompEstimate
   pass-through. Fixed Int dollar zero to Double dollar zero at two call sites.

3. PortfolioIQViewModel.swift: Added gradeCompany/gradeValue to photo
   upload reconstruction site (InventoryCard init was missing the new
   fields).

4. PortfolioDetailPhotosCard.swift: Same fix for photo delete
   reconstruction.

Build succeeded after these fixes. Branch now compiles cleanly.

Bug A (silent BGS 9.5 to nil via Int parsing) and Bug B (decode crash
on decimal grade from backend) from wire-format diagnostic are both
resolved by the Double? alignment.

References:
- 113f091 Windows-side commit that needed these fixes
- 8b4465c backend grade parser + backfill (already on main)
- fbbab52 Phase 1+3 resolver fix (already on main)

+++ HobbyIQ/APIService.swift
@@ -1044,7 +1044,7 @@ struct CardEstimateRequest: Encodable {
     let parallel: String?
     let isAuto: Bool?
     let gradeCompany: String?
-    let gradeValue: Int?
+    let gradeValue: Double?
 
     private enum CodingKeys: String, CodingKey {
         case query
